+++ source/DSCResources/DSC_CMStatusReportingComponent/DSC_CMStatusReportingComponent.psm1
@@ -345,6 +345,7 @@ function Set-TargetResource
 function Test-TargetResource
 {
     [CmdletBinding()]
+    [OutputType([System.Boolean])]
     param
     (
         [Parameter(Mandatory = $true)]