+++ source/Examples/Resources/CMSecurityRoles/CMSecurityRoles_Absent.ps1
@@ -1,0 +1,18 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to remove Security Roles from Configuration Manager.
+#>
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMSecurityRoles ExampleSettings
+        {
+            SiteCode          = 'Lab'
+            SecurityRoleName  = 'Field Services'
+            Ensure            = 'Absent'
+        }
+    }
+}