+++ CHANGELOG.md
@@ -49,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMDistributionPointGroupMembers Resource
 - Added CMSecurityScopes Resource
 - Added CMUserDiscovery Resource
+- Added CMSecurityRoles Resource
 
 ### Changed